patch emacs to respect native-compile-target-directory
Patch batch-native-compile to use the value of native-compile-target-directory
if the optional "for-tarball" parameter is nil.

The previous behavior was to set native-compile-target-directory to
the last directory in the native-comp-eln-load-path list if the optional
"for-tarball" parameter was non-nil, and otherwise set it to nil.

The previous behavior seems to serve no purpose and complicated batch
compiling from ebuilds.